[FIX] website_sale: fix zoom wich was not always computed
In some case, (many JS like runbot with all modules), the onload event was
binded after that the image was loaded. So the JS code that make the zoom
conditionnal was not executed.

Now we apply the function on all images already loaded, or that will is loading

Courtesy of @ged-odoo for help and review