[FIX] sale: missing field in quotation report
Before this commit, the field validity_date was missing from the
default quotation report. Historically this field was added in
module website_quote but it was then moved to sale in
77bd5181f7991de5c3c0fadd1fbada93b6d9d2b1

The modification on the quotation report related to validity_date
were missed at that time.

This commit adds validity_date to the default quotation report.