[FIX] website_sale[_event]: add timout for tour.
Seems like this tour crash randomly due to default 10000 ms.

This button, in some case, seems to take more than 10000ms (slow server? slow
connection ?) to validate the order (cretae invoice, generate PDF, send mail, ...)

It is a test to see if increase the timeout could decrease the risk of false
positive on the runbot.